src/requirements.txt
openai-agents
firecrawl-py
elevenlabs
boto3